[AN] refactor: Search 관련 ViewModel 분리 및 네비게이션 구조 변경 #773
*   [AN] feat: CategoryViewModel, SearchMainViewModel 생성 및 로직 분리
*   [AN] refactor: SearchViewModel의 불필요한 의존성 및 카테고리/메인 관련 로직 제거
*   [AN] refactor: Navigation Type-Safe 적용 및 라우트 정의 변경 (SearchRoute 도입)
*   [AN] refactor: SearchMainRoute, CategoryRoute가 전용 ViewModel을 사용하도록 수정 및 네비게이션 인자 전달 방식 변경
