+++ android/app/src/main/java/com/onair/hearit/presentation/search/main/SearchMainRoute.kt
@@ -3,14 +3,14 @@ package com.onair.hearit.presentation.search.main
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.onair.hearit.presentation.search.SearchViewModel
 
 @Composable
 fun SearchMainRoute(
-    viewModel: SearchViewModel,
     onSearchBarClick: () -> Unit,
     onCategoryClick: (Long, String, String) -> Unit,
+    viewModel: SearchMainViewModel = hiltViewModel(),
 ) {
     val uiState by viewModel.searchMainUiState.collectAsStateWithLifecycle()
